Revert "[AD-976] Implement single-instance SSH Tunnel. (#436)"
