+++ config/spotbugs/spotbugs-exclude.xml
@@ -90,10 +90,7 @@
     <Bug pattern="ODR_OPEN_DATABASE_RESOURCE"/>
   </Match>
   <Match>
-    <Or>
-      <Source name="Driver.java"/>
-      <Source name="DocumentDbSshTunnelServer.java"/>
-    </Or>
+    <Source name="Driver.java"/>
     <Bug pattern="COMMAND_INJECTION"/>
   </Match>
   <Match>
@@ -101,9 +98,6 @@
       <Source name="DocumentDbConnectionProperties.java"/>
       <Source name="DocumentDbConnectionPropertiesTest.java"/>
       <Source name="DocumentDbMain.java"/>
-      <Source name="DocumentDbSshTunnelLock.java"/>
-      <Source name="DocumentDbSshTunnelServer.java"/>
-      <Source name="DocumentDbSshTunnelService.java"/>
     </Or>
     <Bug pattern="PATH_TRAVERSAL_IN"/>
   </Match>
