+++ src/main/java/software/amazon/documentdb/jdbc/DocumentDbConnectionProperties.java
@@ -41,8 +41,6 @@ import java.io.UnsupportedEncodingException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
@@ -52,7 +50,6 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Matcher;
@@ -64,20 +61,19 @@ public class DocumentDbConnectionProperties extends Properties {
     public static final String USER_HOME_PROPERTY = "user.home";
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DocumentDbConnectionProperties.class.getName());
+    private static final String USER_HOME_PATH_NAME  = System.getProperty(USER_HOME_PROPERTY);
+    private static final String DOCUMENTDB_HOME_PATH_NAME = Paths.get(
+            USER_HOME_PATH_NAME, ".documentdb").toString();
     private static final Pattern WHITE_SPACE_PATTERN = Pattern.compile("^\\s*$");
     private static final String ROOT_PEM_RESOURCE_FILE_NAME = "/rds-ca-2019-root.pem";
     public static final String HOME_PATH_PREFIX_REG_EXPR = "^~[/\\\\].*$";
     public static final int FETCH_SIZE_DEFAULT = 2000;
     public static final String DOCUMENTDB_CUSTOM_OPTIONS = "DOCUMENTDB_CUSTOM_OPTIONS";
+    private static String classPathLocationName = null;
     private static String[] documentDbSearchPaths = null;
     private static final String DEFAULT_APPLICATION_NAME_KEY = "default.application.name";
     private static final String PROPERTIES_FILE_PATH = "/documentdb-jdbc.properties";
     static final String DEFAULT_APPLICATION_NAME;
-
-    public static final String USER_HOME_PATH_NAME  = System.getProperty(USER_HOME_PROPERTY);
-    public static final String DOCUMENTDB_HOME_PATH_NAME = Paths.get(
-            USER_HOME_PATH_NAME, ".documentdb").toString();
-    public static final String CONNECTION_STRING_TEMPLATE = "//%s%s/%s%s";
 
     static {
         String defaultAppName = "";
@@ -89,24 +85,6 @@ public class DocumentDbConnectionProperties extends Properties {
             LOGGER.error("Error loading default application name: " + e.getMessage());
         }
         DEFAULT_APPLICATION_NAME = defaultAppName;
-    }
-
-    /**
-     * Enumeration of type of validation.
-     */
-    public enum ValidationType {
-        /**
-         * No validation.
-         */
-        NONE,
-        /**
-         * Validate client connection required properties.
-         */
-        CLIENT,
-        /**
-         * Validate SSH tunnel required properties.
-         */
-        SSH_TUNNEL,
     }
 
     /**
@@ -143,11 +121,36 @@ public class DocumentDbConnectionProperties extends Properties {
     }
 
     /**
-     * Gets the parent folder location of the current class.
+     * Gets the user's home path name.
      *
-     * @return a string representing the parent folder location of the current class.
+     * @return the user's home path name.
      */
-    public static String getClassPathLocation() {
+    static String getUserHomePathName() {
+        return USER_HOME_PATH_NAME;
+    }
+
+    /**
+     * Gets the ~/.documentdb path name.
+     *
+     * @return the ~/.documentdb path name.
+     */
+    static String getDocumentdbHomePathName() {
+        return DOCUMENTDB_HOME_PATH_NAME;
+    }
+
+    /**
+     * Gets the class path's location name.
+     *
+     * @return the class path's location name.
+     */
+    static String getClassPathLocationName() {
+        if (classPathLocationName == null) {
+            classPathLocationName = getClassPathLocation();
+        }
+        return classPathLocationName;
+    }
+
+    private static String getClassPathLocation() {
         String classPathLocation = null;
         final URL classPathLocationUrl = DocumentDbConnectionProperties.class
                 .getProtectionDomain()
@@ -768,151 +771,46 @@ public class DocumentDbConnectionProperties extends Properties {
      *
      * @return a {@link String} with the sanitized connection properties.
      */
-    public @NonNull String buildSanitizedConnectionString() {
-        final String loginInfo = buildLoginInfo(getUser(), null);
-        final String hostInfo = buildHostInfo(getHostname());
-        final String databaseInfo = buildDatabaseInfo(getDatabase());
+    public String buildSanitizedConnectionString() {
+        final String connectionStringTemplate = "//%s%s/%s%s";
+        final String loginInfo = isNullOrWhitespace(getUser()) ? "" : getUser() + "@";
+        final String hostInfo = isNullOrWhitespace(getHostname()) ? "" : getHostname();
+        final String databaseInfo = isNullOrWhitespace(getDatabase()) ? "" : getDatabase();
         final StringBuilder optionalInfo = new StringBuilder();
-        buildSanitizedOptionalInfo(optionalInfo, this);
-        return buildConnectionString(loginInfo, hostInfo, databaseInfo, optionalInfo.toString());
-    }
-
-    @NonNull static String buildDatabaseInfo(final @Nullable String database) {
-        return isNullOrWhitespace(database) ? "" : encodeValue(database);
-    }
-
-    @NonNull static String buildHostInfo(final @Nullable String hostname) {
-        return isNullOrWhitespace(hostname) ? "" : hostname;
-    }
-
-    @NonNull static String buildLoginInfo(final @Nullable String user, final @Nullable String password) {
-        final String userString = isNullOrWhitespace(user)
-                ? ""
-                : encodeValue(user);
-        final String passwordString = isNullOrWhitespace(password)
-                ? ""
-                : ":" + encodeValue(password);
-        final String userInfo = isNullOrWhitespace(userString) && isNullOrWhitespace(passwordString)
-                ? ""
-                : "@";
-        return userString + passwordString + userInfo;
-    }
-
-    static @NonNull String buildConnectionString(
-            final String loginInfo,
-            final String hostInfo,
-            final String databaseInfo,
-            final String optionalInfo) {
-        return String.format(CONNECTION_STRING_TEMPLATE,
-                loginInfo,
-                hostInfo,
-                databaseInfo,
-                optionalInfo);
-    }
-
-    /**
-     * Builds the sanitized optional info connection string. I does not include
-     * sensitive options like SSH_PRIVATE_KEY_PASSPHRASE.
-     *
-     * @param optionalInfo the connection string to build.
-     */
-    static void buildSanitizedOptionalInfo(
-            final StringBuilder optionalInfo,
-            final DocumentDbConnectionProperties properties) {
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.APPLICATION_NAME, properties.getApplicationName());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.LOGIN_TIMEOUT_SEC, properties.getLoginTimeout());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.METADATA_SCAN_LIMIT, properties.getMetadataScanLimit());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.METADATA_SCAN_METHOD, properties.getMetadataScanMethod());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.RETRY_READS_ENABLED, properties.getRetryReadsEnabled());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.READ_PREFERENCE, properties.getReadPreference());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.REPLICA_SET, properties.getReplicaSet(), null);
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.TLS_ENABLED, properties.getTlsEnabled());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.TLS_ALLOW_INVALID_HOSTNAMES, properties.getTlsAllowInvalidHostnames());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.TLS_CA_FILE, properties.getTlsCAFilePath(), null);
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.SCHEMA_NAME, properties.getSchemaName());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.SSH_USER, properties.getSshUser(), null);
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.SSH_HOSTNAME, properties.getSshHostname(), null);
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.SSH_PRIVATE_KEY_FILE, properties.getSshPrivateKeyFile(), null);
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.SSH_STRICT_HOST_KEY_CHECKING, properties.getSshStrictHostKeyChecking());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.SSH_KNOWN_HOSTS_FILE, properties.getSshKnownHostsFile(), null);
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.DEFAULT_FETCH_SIZE, properties.getDefaultFetchSize());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.REFRESH_SCHEMA, properties.getRefreshSchema());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.DEFAULT_AUTH_DB, properties.getDefaultAuthenticationDatabase());
-        maybeAppendOptionalValue(optionalInfo, DocumentDbConnectionProperty.ALLOW_DISK_USE, properties.getAllowDiskUseOption());
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final String value) {
-        if (!property.getDefaultValue().equals(value)) {
-            appendOption(optionalInfo, property, value);
+        if (!getApplicationName()
+                .equals(DocumentDbConnectionProperty.APPLICATION_NAME.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.APPLICATION_NAME, getApplicationName());
         }
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final String value,
-                                  final String defaultValue) {
-        if (!Objects.equals(defaultValue, value)) {
-            appendOption(optionalInfo, property, value);
+        if (getLoginTimeout() != Integer.parseInt(DocumentDbConnectionProperty.LOGIN_TIMEOUT_SEC.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.LOGIN_TIMEOUT_SEC, getLoginTimeout());
         }
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final int value) {
-        if (value != Integer.parseInt(property.getDefaultValue())) {
-            appendOption(optionalInfo, property, value);
+        if (getMetadataScanLimit() != Integer.parseInt(DocumentDbConnectionProperty.METADATA_SCAN_LIMIT.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.METADATA_SCAN_LIMIT, getMetadataScanLimit());
         }
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final boolean value) {
-        if (value != Boolean.parseBoolean(property.getDefaultValue())) {
-            appendOption(optionalInfo, property, value);
+        if (getMetadataScanMethod() != DocumentDbMetadataScanMethod.fromString(DocumentDbConnectionProperty.METADATA_SCAN_METHOD.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.METADATA_SCAN_METHOD, getMetadataScanMethod().getName());
         }
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final DocumentDbMetadataScanMethod value) {
-        if (value != DocumentDbMetadataScanMethod.fromString(property.getDefaultValue())) {
-            appendOption(optionalInfo, property, value.getName());
+        if (getRetryReadsEnabled() != Boolean.parseBoolean(DocumentDbConnectionProperty.RETRY_READS_ENABLED.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.RETRY_READS_ENABLED, getRetryReadsEnabled());
         }
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final DocumentDbReadPreference value) {
-        if (value != null) {
-            appendOption(optionalInfo, property, value.getName());
+        if (getReadPreference() != null) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.READ_PREFERENCE, getReadPreference().getName());
         }
-    }
-
-    static void maybeAppendOptionalValue(final StringBuilder optionalInfo,
-                                  final DocumentDbConnectionProperty property,
-                                  final DocumentDbAllowDiskUseOption value) {
-        if (value != DocumentDbAllowDiskUseOption.fromString(property.getDefaultValue())) {
-            appendOption(optionalInfo, property, value.getName());
+        if (getReplicaSet() != null) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.REPLICA_SET, getReplicaSet());
         }
-    }
-
-    /**
-     * Builds the connection string for SSH properties.
-     *
-     * @return a connection string with SSH properties.
-     */
-    public String buildSshConnectionString() {
-        final String loginInfo = "";
-        final String hostInfo = buildHostInfo(getHostname());
-        final String databaseInfo = "";
-        final StringBuilder optionalInfo = new StringBuilder();
-        buildSshOptionalInfo(optionalInfo);
-        return buildConnectionString(loginInfo, hostInfo, databaseInfo, optionalInfo.toString());
-    }
-
-    private void buildSshOptionalInfo(final StringBuilder optionalInfo) {
+        if (getTlsEnabled() != Boolean.parseBoolean(DocumentDbConnectionProperty.TLS_ENABLED.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.TLS_ENABLED, getTlsEnabled());
+        }
+        if (getTlsAllowInvalidHostnames() != Boolean.parseBoolean(DocumentDbConnectionProperty.TLS_ALLOW_INVALID_HOSTNAMES.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.TLS_ALLOW_INVALID_HOSTNAMES, getTlsAllowInvalidHostnames());
+        }
+        if (getTlsCAFilePath() != null) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.TLS_CA_FILE, getTlsCAFilePath());
+        }
+        if (!DocumentDbConnectionProperty.SCHEMA_NAME.getDefaultValue().equals(getSchemaName())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.SCHEMA_NAME, getSchemaName());
+        }
         if (getSshUser() != null) {
             appendOption(optionalInfo, DocumentDbConnectionProperty.SSH_USER, getSshUser());
         }
@@ -922,45 +820,36 @@ public class DocumentDbConnectionProperties extends Properties {
         if (getSshPrivateKeyFile() != null) {
             appendOption(optionalInfo, DocumentDbConnectionProperty.SSH_PRIVATE_KEY_FILE, getSshPrivateKeyFile());
         }
-        if (getSshPrivateKeyPassphrase() != null && !DocumentDbConnectionProperty.SSH_PRIVATE_KEY_PASSPHRASE.getDefaultValue().equals(getSshPrivateKeyPassphrase())) {
-            appendOption(optionalInfo, DocumentDbConnectionProperty.SSH_PRIVATE_KEY_PASSPHRASE, getSshPrivateKeyPassphrase());
-        }
         if (getSshStrictHostKeyChecking() != Boolean.parseBoolean(DocumentDbConnectionProperty.SSH_STRICT_HOST_KEY_CHECKING.getDefaultValue())) {
             appendOption(optionalInfo, DocumentDbConnectionProperty.SSH_STRICT_HOST_KEY_CHECKING, getSshStrictHostKeyChecking());
         }
         if (getSshKnownHostsFile() != null && !DocumentDbConnectionProperty.SSH_KNOWN_HOSTS_FILE.getDefaultValue().equals(getSshKnownHostsFile())) {
             appendOption(optionalInfo, DocumentDbConnectionProperty.SSH_KNOWN_HOSTS_FILE, getSshKnownHostsFile());
         }
+        if (getDefaultFetchSize() != Integer.parseInt(DocumentDbConnectionProperty.DEFAULT_FETCH_SIZE.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.DEFAULT_FETCH_SIZE, getDefaultFetchSize());
+        }
+        if (getRefreshSchema() != Boolean.parseBoolean(DocumentDbConnectionProperty.REFRESH_SCHEMA.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.REFRESH_SCHEMA, getRefreshSchema());
+        }
+        if (getDefaultAuthenticationDatabase() != null && !DocumentDbConnectionProperty.DEFAULT_AUTH_DB.getDefaultValue().equals(getDefaultAuthenticationDatabase())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.DEFAULT_AUTH_DB, getDefaultAuthenticationDatabase());
+        }
+        if (getAllowDiskUseOption() != DocumentDbAllowDiskUseOption.fromString(DocumentDbConnectionProperty.ALLOW_DISK_USE.getDefaultValue())) {
+            appendOption(optionalInfo, DocumentDbConnectionProperty.ALLOW_DISK_USE, getAllowDiskUseOption().getName());
+        }
+        return String.format(connectionStringTemplate,
+                loginInfo,
+                hostInfo,
+                databaseInfo,
+                optionalInfo);
     }
 
-    /**
-     * Appends an option and value to the string.
-     *
-     * @param optionInfo the connection string to build.
-     * @param option the option to add.
-     * @param optionValue the option value to set.
-     */
-    public static void appendOption(final StringBuilder optionInfo,
+    private void appendOption(final StringBuilder optionInfo,
             final DocumentDbConnectionProperty option,
             final Object optionValue) {
         optionInfo.append(optionInfo.length() == 0 ? "?" : "&");
-        optionInfo.append(option.getName())
-                .append("=")
-                .append(optionValue == null ? "" : encodeValue(optionValue.toString()));
-    }
-
-    /**
-     * Encodes a value into URL encoded value.
-     *
-     * @param value the value to encode.
-     * @return the encoded value.
-     */
-    public static String encodeValue(final String value) {
-        try {
-            return URLEncoder.encode(value, StandardCharsets.UTF_8.toString());
-        } catch (UnsupportedEncodingException e) {
-            return value;
-        }
+        optionInfo.append(option.getName()).append("=").append(optionValue);
     }
 
     /**
@@ -968,58 +857,26 @@ public class DocumentDbConnectionProperties extends Properties {
      * @throws SQLException if the required properties are not correctly set.
      */
     public void validateRequiredProperties() throws SQLException {
-        validateRequiredProperties(ValidationType.CLIENT);
-    }
-
-    /**
-     * Validates the existing properties.
-     * @param validationType Which validation type to perform.
-     * @throws SQLException if the required properties are not correctly set.
-     */
-    public void validateRequiredProperties(final ValidationType validationType) throws SQLException {
-        if ((isNullOrWhitespace(getUser())
-                || isNullOrWhitespace(getPassword())) && validationType == ValidationType.CLIENT) {
+        if (isNullOrWhitespace(getUser())
+                || isNullOrWhitespace(getPassword())) {
             throw SqlError.createSQLException(
                     LOGGER,
                     SqlState.INVALID_PARAMETER_VALUE,
                     SqlError.MISSING_USER_PASSWORD
             );
         }
-        if (isNullOrWhitespace(getDatabase()) && validationType == ValidationType.CLIENT) {
+        if (isNullOrWhitespace(getDatabase())) {
             throw SqlError.createSQLException(
                     LOGGER,
                     SqlState.INVALID_PARAMETER_VALUE,
                     SqlError.MISSING_DATABASE
             );
         }
-        if (isNullOrWhitespace(getHostname())
-                && (validationType == ValidationType.CLIENT || validationType == ValidationType.SSH_TUNNEL)) {
+        if (isNullOrWhitespace(getHostname())) {
             throw SqlError.createSQLException(
                     LOGGER,
                     SqlState.INVALID_PARAMETER_VALUE,
                     SqlError.MISSING_HOSTNAME
-            );
-        }
-
-        if (isNullOrWhitespace(getSshUser()) && validationType == ValidationType.SSH_TUNNEL) {
-            throw SqlError.createSQLException(
-                    LOGGER,
-                    SqlState.INVALID_PARAMETER_VALUE,
-                    SqlError.MISSING_SSH_USER
-            );
-        }
-        if (isNullOrWhitespace(getSshHostname()) && validationType == ValidationType.SSH_TUNNEL) {
-            throw SqlError.createSQLException(
-                    LOGGER,
-                    SqlState.INVALID_PARAMETER_VALUE,
-                    SqlError.MISSING_SSH_HOSTNAME
-            );
-        }
-        if (isNullOrWhitespace(getSshPrivateKeyFile()) && validationType == ValidationType.SSH_TUNNEL) {
-            throw SqlError.createSQLException(
-                    LOGGER,
-                    SqlState.INVALID_PARAMETER_VALUE,
-                    SqlError.MISSING_SSH_PRIVATE_KEY_FILE
             );
         }
     }
@@ -1036,19 +893,6 @@ public class DocumentDbConnectionProperties extends Properties {
         return getPropertiesFromConnectionString(new Properties(), documentDbUrl, DOCUMENT_DB_SCHEME);
     }
 
-    /**
-     * Gets the connection properties from the connection string.
-     *
-     * @param documentDbUrl the given properties.
-     * @param validationType Which properties to validate.
-     * @return a {@link DocumentDbConnectionProperties} with the properties set.
-     * @throws SQLException if connection string is invalid.
-     */
-    public static DocumentDbConnectionProperties getPropertiesFromConnectionString(
-            final String documentDbUrl,
-            final ValidationType validationType) throws SQLException {
-        return getPropertiesFromConnectionString(new Properties(), documentDbUrl, DOCUMENT_DB_SCHEME, validationType);
-    }
 
     /**
      * Gets the connection properties from the connection string.
@@ -1062,22 +906,6 @@ public class DocumentDbConnectionProperties extends Properties {
     public static DocumentDbConnectionProperties getPropertiesFromConnectionString(
             final Properties info, final String documentDbUrl, final String connectionStringPrefix)
             throws SQLException {
-        return getPropertiesFromConnectionString(info, documentDbUrl, connectionStringPrefix, ValidationType.CLIENT);
-    }
-
-    /**
-     * Gets the connection properties from the connection string.
-     *
-     * @param info the given properties.
-     * @param documentDbUrl the connection string.
-     * @param connectionStringPrefix the connection string prefix.
-     * @param validationType Which validation to perform.
-     * @return a {@link DocumentDbConnectionProperties} with the properties set.
-     * @throws SQLException if connection string is invalid.
-     */
-    public static DocumentDbConnectionProperties getPropertiesFromConnectionString(
-            final Properties info, final String documentDbUrl, final String connectionStringPrefix,
-            final ValidationType validationType) throws SQLException {
 
         final DocumentDbConnectionProperties properties = new DocumentDbConnectionProperties(info);
         final String postSchemeSuffix = documentDbUrl.substring(connectionStringPrefix.length());
@@ -1085,11 +913,11 @@ public class DocumentDbConnectionProperties extends Properties {
             try {
                 final URI uri = new URI(postSchemeSuffix);
 
-                setHostName(properties, uri, validationType);
+                setHostName(properties, uri);
 
-                setUserPassword(properties, uri, validationType);
+                setUserPassword(properties, uri);
 
-                setDatabase(properties, uri, validationType);
+                setDatabase(properties, uri);
 
                 setOptionalProperties(properties, uri);
 
@@ -1099,17 +927,15 @@ public class DocumentDbConnectionProperties extends Properties {
                 throw SqlError.createSQLException(
                         LOGGER,
                         SqlState.CONNECTION_FAILURE,
-                        e,
                         SqlError.INVALID_CONNECTION_PROPERTIES,
-                        documentDbUrl + " : '" + e.getMessage() + "'"
+                        documentDbUrl
                 );
             } catch (UnsupportedEncodingException e) {
                 throw new SQLException(e.getMessage(), e);
             }
         }
 
-        properties.validateRequiredProperties(validationType);
-
+        properties.validateRequiredProperties();
         return properties;
     }
 
@@ -1128,13 +954,11 @@ public class DocumentDbConnectionProperties extends Properties {
         }
     }
 
-    private static void setDatabase(
-            final Properties properties,
-            final URI mongoUri,
-            final ValidationType validationType) throws SQLException {
+    private static void setDatabase(final Properties properties, final URI mongoUri)
+            throws SQLException {
         if (isNullOrWhitespace(mongoUri.getPath())) {
-            if (properties.getProperty(DocumentDbConnectionProperty.DATABASE.getName(), null) == null
-                    && validationType == ValidationType.CLIENT) {
+            if (properties.getProperty(
+                    DocumentDbConnectionProperty.DATABASE.getName(), null) == null) {
                 throw SqlError.createSQLException(
                         LOGGER,
                         SqlState.CONNECTION_FAILURE,
@@ -1164,15 +988,13 @@ public class DocumentDbConnectionProperties extends Properties {
         }
     }
 
-    private static void setUserPassword(
-            final Properties properties,
-            final URI mongoUri,
-            final ValidationType validationType) throws UnsupportedEncodingException, SQLException {
+    private static void setUserPassword(final Properties properties, final URI mongoUri)
+            throws UnsupportedEncodingException, SQLException {
         if (mongoUri.getUserInfo() == null) {
-            if ((properties.getProperty(DocumentDbConnectionProperty.USER.getName(), null) == null
-                    && validationType == ValidationType.CLIENT)
-                    || (properties.getProperty(DocumentDbConnectionProperty.PASSWORD.getName(), null) == null
-                    && validationType == ValidationType.CLIENT)) {
+            if (properties.getProperty(
+                    DocumentDbConnectionProperty.USER.getName(), null) == null
+                    || properties.getProperty(
+                    DocumentDbConnectionProperty.PASSWORD.getName(), null) == null) {
                 throw SqlError.createSQLException(
                         LOGGER,
                         SqlState.CONNECTION_FAILURE,
@@ -1196,14 +1018,11 @@ public class DocumentDbConnectionProperties extends Properties {
         }
     }
 
-    private static void setHostName(
-            final Properties properties,
-            final URI mongoUri,
-            final ValidationType validationType) throws SQLException {
+    private static void setHostName(final Properties properties, final URI mongoUri) throws SQLException {
         String hostName = mongoUri.getHost();
         if (hostName == null) {
-            if (properties.getProperty(DocumentDbConnectionProperty.HOSTNAME.getName(), null) == null
-                    && (validationType == ValidationType.CLIENT || validationType == ValidationType.SSH_TUNNEL)) {
+            if (properties.getProperty(
+                    DocumentDbConnectionProperty.HOSTNAME.getName(), null) == null) {
                 throw SqlError.createSQLException(
                         LOGGER,
                         SqlState.CONNECTION_FAILURE,
@@ -1410,31 +1229,31 @@ public class DocumentDbConnectionProperties extends Properties {
      * Gets an absolute path from the given file path. It performs the substitution for a leading
      * '~' to be replaced by the user's home directory.
      *
-     * @param filePathString the given file path to process.
+     * @param filePath the given file path to process.
      * @param searchFolders list of folders
      * @return a {@link Path} for the absolution path for the given file path.
      */
-    public static Path getPath(final String filePathString, final String... searchFolders) {
-        final Path filePath = Paths.get(filePathString);
-        if (filePathString.matches(HOME_PATH_PREFIX_REG_EXPR)) {
-            final String fromHomePath = filePathString.replaceFirst("~",
+    public static Path getPath(final String filePath, final String... searchFolders) {
+        if (filePath.matches(HOME_PATH_PREFIX_REG_EXPR)) {
+            final String fromHomePath = filePath.replaceFirst("~",
                     Matcher.quoteReplacement(USER_HOME_PATH_NAME));
             return Paths.get(fromHomePath).toAbsolutePath();
         } else {
-            if (filePath.isAbsolute()) {
-                return filePath;
+            final Path origFilePath = Paths.get(filePath);
+            if (origFilePath.isAbsolute()) {
+                return origFilePath;
             }
             for (String searchFolder : searchFolders) {
                 if (searchFolder == null) {
                     continue;
                 }
-                final Path testPath = Paths.get(searchFolder, filePathString);
+                final Path testPath = Paths.get(searchFolder, filePath);
                 if (testPath.toAbsolutePath().toFile().exists()) {
                     return testPath;
                 }
             }
         }
-        return filePath.toAbsolutePath();
+        return Paths.get(filePath).toAbsolutePath();
     }
 
     /**
@@ -1469,7 +1288,7 @@ public class DocumentDbConnectionProperties extends Properties {
             if (getProperty(key) != null) {
                 property = DocumentDbMetadataScanMethod.fromString(getProperty(key));
             } else if (DocumentDbConnectionProperty.getPropertyFromKey(key) != null) {
-                property = DocumentDbMetadataScanMethod.fromString(
+                property =  DocumentDbMetadataScanMethod.fromString(
                         DocumentDbConnectionProperty.getPropertyFromKey(key).getDefaultValue());
             }
         } catch (IllegalArgumentException e) {
