+++ src/main/java/software/amazon/documentdb/jdbc/common/utilities/SqlError.java
@@ -55,12 +55,8 @@ public enum SqlError {
     KNOWN_HOSTS_FILE_NOT_FOUND,
     MISSING_DATABASE,
     MISSING_HOSTNAME,
-    MISSING_JAVA_HOME,
     MISSING_SCHEMA,
     MISSING_PASSWORD,
-    MISSING_SSH_USER,
-    MISSING_SSH_HOSTNAME,
-    MISSING_SSH_PRIVATE_KEY_FILE,
     MISSING_USER_PASSWORD,
     MISSING_LITERAL_VALUE,
     MISMATCH_SCHEMA_NAME,
@@ -77,10 +73,8 @@ public enum SqlError {
     RESULT_SET_CLOSED,
     SECURITY_ERROR,
     SSH_PRIVATE_KEY_FILE_NOT_FOUND,
-    SSH_TUNNEL_PATH_NOT_FOUND,
     SINGLE_EQUIJOIN_ONLY,
     SQL_PARSE_ERROR,
-    SSH_TUNNEL_ERROR,
     STMT_CLOSED,
     TLS_CA_FILE_NOT_FOUND,
     TRANSACTIONS_NOT_SUPPORTED,
@@ -120,10 +114,7 @@ public enum SqlError {
      * @return resource String, formatted with formatArgs.
      */
     public static String lookup(final SqlError key, final Object... formatArgs) {
-        // Remove any new lines.
-        return String
-                .format(RESOURCE.getString(key.name()), formatArgs)
-                .replaceAll("[\r\n]", "");
+        return String.format(RESOURCE.getString(key.name()), formatArgs);
     }
 
     /**
