+++ src/main/resources/jdbc.properties
@@ -30,8 +30,6 @@ READ_ONLY=Driver is read-only. Please refer to the list of the supported feature
 RESULT_FORWARD_ONLY=Cannot retrieve previous rows. Please refer to the list of the supported features.
 RESULT_SET_CLOSED=Result set is closed.
 SSH_PRIVATE_KEY_FILE_NOT_FOUND=SSH private key file '%s' not found.
-SSH_TUNNEL_PATH_NOT_FOUND=Path to SSH Tunnel '%s' not found or not executable.
-SSH_TUNNEL_ERROR=Error reported from SSH Tunnel service. (%s)
 STMT_CLOSED=Statement is closed.
 TRANSACTIONS_NOT_SUPPORTED=Transactions are not supported.
 VALUE_TRUNCATED=The value of type %s has been truncated while converting to the target type %s.
@@ -66,7 +64,7 @@ INVALID_STATE_SET_TABLE_FUNCTION=This method must only be called once after load
 UPSERT_SCHEMA_FAILED=Failed to insert or update one or more records for schema '%s'. Please report the issue to AWS support.
 
 ### Typical Error Messages (may occur in BI tools)
-SQL_PARSE_ERROR=Unable to parse SQL '%s'. -- Reason: '%s'
+SQL_PARSE_ERROR=Unable to parse SQL '%s'.%n Reason: '%s'
 QUERY_FAILED=Execute query has failed, error: %s.
 CONN_CLOSED=Connection has been closed. Please reconnect.
 CREATE_FOLDER_FAILED=Unable to create folder '%s'. Check file path and permissions.
@@ -78,12 +76,8 @@ INVALID_CONNECTION_PROPERTIES=Invalid JDBC connection properties in url: '%s'. C
 INVALID_FORMAT=Value '%s' has invalid format. Expecting format '%s'. Please regenerate the schema.
 MISSING_DATABASE=Database is required to connect. Syntax: 'jdbc:documentdb://[<user>[:<password>]@]<hostname>/<database>[?options...]'
 MISSING_HOSTNAME=Valid hostname is required to connect. Syntax: 'jdbc:documentdb://[<user>[:<password>]@]<hostname>/<database>[?options...]'
-MISSING_JAVA_HOME=The environment variable JAVA_HOME is not set or is invalid.
 MISSING_SCHEMA=The schema '%s' is not found.
 MISSING_PASSWORD=Password must not be empty.
-MISSING_SSH_USER=SSH Tunnel user must not be empty.
-MISSING_SSH_HOSTNAME=SSH Tunnel hostname must not be empty.
-MISSING_SSH_PRIVATE_KEY_FILE=SSH Tunnel private key filename must not be empty.
 MISSING_USER_PASSWORD=User and password are required to connect. Syntax: 'jdbc:documentdb://[<user>[:<password>]@]<hostname>/<database>[?options...]'
 MISSING_LITERAL_VALUE=Literal value of type %s is unexpectedly missing.
 MISMATCH_SCHEMA_NAME=Given schema name '%s' does not match stored schema name '%s'.
