+++ src/testFixtures/java/software/amazon/documentdb/jdbc/common/test/DocumentDbAbstractTestEnvironment.java
@@ -43,6 +43,9 @@ import software.amazon.documentdb.jdbc.DocumentDbConnectionProperties;
 import software.amazon.documentdb.jdbc.DocumentDbConnectionProperty;
 import software.amazon.documentdb.jdbc.DocumentDbMetadataScanMethod;
 
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.sql.SQLException;
 import java.time.Instant;
 import java.util.AbstractMap.SimpleEntry;
@@ -50,8 +53,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map.Entry;
 import java.util.UUID;
-
-import static software.amazon.documentdb.jdbc.DocumentDbConnectionProperties.encodeValue;
 
 /**
  * Provides some base implementation of the {@link DocumentDbTestEnvironment} interface.
@@ -168,12 +169,9 @@ public abstract class DocumentDbAbstractTestEnvironment implements DocumentDbTes
 
     @Override
     public boolean start() throws Exception {
-        if (!isStarted) {
-            final boolean started = startEnvironment();
-            isStarted = true;
-            return started;
-        }
-        return false;
+        final boolean started = startEnvironment();
+        isStarted = true;
+        return started;
     }
 
     @Override
@@ -268,6 +266,14 @@ public abstract class DocumentDbAbstractTestEnvironment implements DocumentDbTes
     private String getCredentials(final boolean isRestrictedUser) {
         return username != null && password != null
                 ? String.format("%s:%s@", encodeValue(isRestrictedUser ? restrictedUsername : username), encodeValue(password)) : "";
+    }
+
+    private static String encodeValue(final String value) {
+        try {
+            return URLEncoder.encode(value, StandardCharsets.UTF_8.toString());
+        } catch (UnsupportedEncodingException e) {
+            return value;
+        }
     }
 
     @Override
